[volume-7] Decoupling with Event

+++ apps/commerce-api/src/main/java/com/loopers/application/user/UserInfo.java
@@ -9,7 +9,7 @@ public record UserInfo(
         String email,
         String birthDate,
         String gender,
-        Integer point) {
+        Long point) {
     public static UserInfo from(UserModel model) {
         return new UserInfo(
             model.getUserId(),
